Move communicator date response to its own hook
