+++ src/components/safe-apps/AppFrame/index.tsx
@@ -1,14 +1,6 @@
 import { ReactElement, useCallback, useEffect, useMemo } from 'react'
 import { CircularProgress, Typography } from '@mui/material'
-import { getAddress } from 'ethers/lib/utils'
-import { getBalances, getTransactionDetails, SafeAppData } from '@gnosis.pm/safe-react-gateway-sdk'
-import {
-  GetBalanceParams,
-  GetTxBySafeTxHashParams,
-  Methods,
-  SendTransactionsParams,
-  SignMessageParams,
-} from '@gnosis.pm/safe-apps-sdk'
+import { SafeAppData } from '@gnosis.pm/safe-react-gateway-sdk'
 import { trackSafeAppOpenCount } from '@/services/safe-apps/track-app-usage-count'
 import { useSafeAppFromManifest } from '@/hooks/safe-apps/useSafeAppFromManifest'
 import useSafeInfo from '@/hooks/useSafeInfo'
@@ -17,20 +9,9 @@ import { isSameUrl } from '@/utils/url'
 import { ThirdPartyCookiesWarning } from './ThirdPartyCookiesWarning'
 import useThirdPartyCookies from './useThirdPartyCookies'
 import useAppIsLoading from './useAppIsLoading'
-import useAppCommunicator from './useAppCommunicator'
-import { useCurrentChain } from '@/hooks/useChains'
-import useIsGranted from '@/hooks/useIsGranted'
-import useSafeAddress from '@/hooks/useSafeAddress'
-import { createSafeAppsWeb3Provider } from '@/hooks/wallets/web3'
 
 import css from './styles.module.css'
-
-type JsonRpcResponse = {
-  jsonrpc: string
-  id: number
-  result?: any
-  error?: string
-}
+import useAppCommunicator from './useAppCommunicator'
 
 type AppFrameProps = {
   appUrl: string
@@ -38,23 +19,13 @@ type AppFrameProps = {
 
 const AppFrame = ({ appUrl }: AppFrameProps): ReactElement => {
   const { safe } = useSafeInfo()
-  const safeAddress = useSafeAddress()
-  const chain = useCurrentChain()
-  const { nativeCurrency, chainName, chainId, shortName, blockExplorerUriTemplate } = chain || { chainId: '' }
+
   const [remoteApps] = useRemoteSafeApps()
   const { safeApp: safeAppFromManifest } = useSafeAppFromManifest(appUrl, safe.chainId)
   const { thirdPartyCookiesDisabled, setThirdPartyCookiesDisabled } = useThirdPartyCookies()
   const { iframeRef, appIsLoading, isLoadingSlow, setAppIsLoading } = useAppIsLoading()
-  const communicator = useAppCommunicator(iframeRef, safeAppFromManifest)
-  const granted = useIsGranted()
 
-  const safeAppWeb3Provider = useMemo(() => {
-    if (!chain) {
-      return
-    }
-
-    return createSafeAppsWeb3Provider(chain)
-  }, [chain])
+  useAppCommunicator(iframeRef, safeAppFromManifest)
 
   const remoteApp = useMemo(() => remoteApps?.find((app: SafeAppData) => app.url === appUrl), [remoteApps, appUrl])
 
@@ -72,75 +43,6 @@ const AppFrame = ({ appUrl }: AppFrameProps): ReactElement => {
 
     setAppIsLoading(false)
   }, [appUrl, iframeRef, setAppIsLoading])
-
-  useEffect(() => {
-    communicator?.on(Methods.getTxBySafeTxHash, async (msg) => {
-      const { safeTxHash } = msg.data.params as GetTxBySafeTxHashParams
-
-      const tx = await getTransactionDetails(chainId, safeTxHash)
-
-      return tx
-    })
-
-    communicator?.on(Methods.getEnvironmentInfo, async () => ({
-      origin: document.location.origin,
-    }))
-
-    communicator?.on(Methods.getSafeInfo, () => ({
-      safeAddress,
-      chainId: parseInt(chainId, 10),
-      owners: safe.owners.map((owner) => owner.value),
-      threshold: safe.threshold,
-      isReadOnly: !granted,
-    }))
-
-    communicator?.on(Methods.getSafeBalances, async (msg) => {
-      const { currency = 'usd' } = msg.data.params as GetBalanceParams
-
-      return getBalances(chainId, safeAddress, currency, {
-        exclude_spam: true,
-        trusted: false,
-      })
-    })
-
-    communicator?.on(Methods.sendTransactions, (msg) => {
-      const { txs, params } = msg.data.params as SendTransactionsParams
-
-      const transactions = txs.map(({ to, ...rest }) => ({
-        to: getAddress(to),
-        ...rest,
-      }))
-
-      // openConfirmationModal(transactions, params, msg.data.id)
-    })
-
-    communicator?.on(Methods.signMessage, async (msg) => {
-      const { message } = msg.data.params as SignMessageParams
-
-      // openSignMessageModal(message, msg.data.id)
-    })
-
-    communicator?.on(Methods.getChainInfo, async () => {
-      return {
-        chainName,
-        chainId,
-        shortName,
-        nativeCurrency,
-        blockExplorerUriTemplate,
-      }
-    })
-  }, [
-    blockExplorerUriTemplate,
-    chainId,
-    chainName,
-    communicator,
-    granted,
-    nativeCurrency,
-    safe,
-    safeAddress,
-    safeAppWeb3Provider,
-    shortName,
-  ])
 
   return (
     <div className={css.wrapper}>
